# perf: no len if it is not needed