Check for existence, not `len`.