@tinyb0t 2023-08-21
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/assembly.py               1               1
tinygrad/codegen/assembly_arm64.py         1               1
tinygrad/codegen/linearizer.py             2               2
tinygrad/codegen/optimizer.py              5               5
tinygrad/lazy.py                           2               2
tinygrad/renderer/cstyle.py                1               1
tinygrad/renderer/wgsl.py                  2               2
tinygrad/runtime/ops_gpu.py                1               1
tinygrad/shape/shapetracker.py             1               1
tinygrad/shape/symbolic.py                 1               1
tinygrad/tensor.py                         1               1
------------------------------------------------------------
total                                     18              18
------------------------------------------------------------
lines added in the tinygrad folder: 0
```