@geohot 2023-08-21
Is this a measurable difference? If it's not, I don't think it's worth merging and risking bugs.