@roelofvandijk 2023-08-21
This is using the fact that [empty built-in types are `False`](https://docs.python.org/3/library/stdtypes.html#truth-value-testing).

[PEP8](https://peps.python.org/pep-0008/) recommends this syntax:

![image](https://github.com/tinygrad/tinygrad/assets/3604013/243f77ca-2838-479b-8e29-2639355ba4df)

See also: https://stackoverflow.com/questions/53513/how-do-i-check-if-a-list-is-empty

Is it faster? Slightly, but hard to measure and within the measurment uncertainty. Each single instance is faster.

The tests say yes:

Before:
![image](https://github.com/tinygrad/tinygrad/assets/3604013/e6ea8829-c063-4e0d-a52c-7f6cbd3ccf89)

After:
![image](https://github.com/tinygrad/tinygrad/assets/3604013/ff9f8fae-52ba-40ff-b22e-2d5f750f3096)


`python3.11 -O test/external/external_test_speed_llama.py` also, albeit a bit mixed.

before:
```
codegen         mean runtime:  189.62ms, runs:   188.22,  243.99,  175.69,  177.16,  176.13,  174.63,  181.40,  198.67,  192.77,  187.58
methodcache     mean runtime:  183.87ms, runs:   219.49,  175.73,  162.25,  218.23,  183.71,  169.01,  172.37,  172.96,  191.82,  173.11
profile         mean runtime:  842.92ms, runs:   789.43,  897.05,  803.74,  847.90,  871.77,  819.39,  829.93,  866.00,  845.42,  858.59

497631/348415  150.925    0.000  218.318    0.001 {built-in method builtins.len}
```

after:
```
codegen         mean runtime:  192.72ms, runs:   218.66,  242.19,  188.81,  182.78,  176.57,  175.60,  184.17,  195.40,  179.69,  183.29
methodcache     mean runtime:  176.60ms, runs:   200.99,  170.58,  164.93,  159.22,  160.60,  165.63,  165.07,  165.50,  182.25,  231.28
profile         mean runtime:  796.60ms, runs:   761.58,  789.26,  863.90,  767.58,  818.63,  785.97,  786.19,  821.58,  782.10,  789.17

478051/361955  135.459    0.000  186.942    0.001 {built-in method builtins.len}
```