@roelofvandijk 2023-08-21
For me, this definitely makes the code more readable, but maybe that's because I am used to it.