@geohot 2023-08-21
I'm sold that this is more readable